requirements.txt
absl-py==0.12.0
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
autoflake==1.4
Babel==2.9.0
backcall==0.2.0
bandit==1.7.0
black==20.8b1
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.5
cudasift==0.1.2
cycler==0.10.0
Cython==0.29.23
darglint==1.8.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
doc8==0.8.1
docutils==0.17
entrypoints==0.3
eradicate==1.0
facenet-pytorch==2.5.2
ffmpeg-python==0.2.0
filelock==3.0.12
flake8==3.9.0
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.4.0
flake8-debugger==3.2.1
flake8-docstrings==1.6.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
freetype-py==2.2.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.1
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==3.10.1
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
mpmath==1.2.1
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nose==1.3.7
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pbr==5.5.1
pep8-naming==0.9.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodomex==3.10.1
pydocstyle==5.1.1
pyflakes==2.3.1
pyglet==1.5.16
Pygments==2.8.1
pygtrie==2.4.2
PyOpenGL==3.1.0
pyparsing==2.4.7
pyrender==0.1.45
pyrsistent==0.17.3
pytest==6.2.3
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dev-tools==2020.9.10
python-ndn==0.3a1.post1
pytz==2021.1
pyupgrade==2.12.0
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
restructuredtext-lint==1.3.2
rsa==4.7.2
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stevedore==3.3.0
sympy==1.8
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.4
testfixtures==6.17.1
testpath==0.4.4
thop==0.0.31.post2005241907
tokenize-rt==4.1.0
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tox==3.23.0
tox-travis==0.12
tqdm==4.60.0
traitlets==5.0.5
trimesh==3.9.15
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wemake-python-styleguide==0.14.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
